Implement servo hardware class for roborio.

+++ cameramount/src/main/java/org/team997coders/spartanlib/interfaces/IServo.java
@@ -6,9 +6,31 @@ package org.team997coders.spartanlib.interfaces;
  * adapted to other Servo classes? We will see...
  */
 public interface IServo {
+  /**
+   * Attach the given pin to the next free channel, sets pinMode, returns channel number or 0 if failure
+   * @param pinArg  The pin to which the servo is attached
+   * @return        The PWM channel attached to or zero if failure
+   */
   byte attach(int pinArg);
+  /**
+   * Returns true if there is a servo attached. 
+   * @return  True if attached.
+   */
   byte attached();
+  /**
+   * Stops an attached servos from pulsing its i/o pin.
+   */
   void detach();
-  byte read();
+  /**
+   * Gets the last written servo pulse width as an angle.
+   * @return   Returns an angle between 0 and 180.
+   */
+  int read();
+  /**
+   * Sets the servo angle in degrees.
+   * Some libraries, with an invalid angle that is valid as pulse in microseconds will be treated as microseconds.
+   * If not valid, an implementaiton should throw an IllegalArgumentException exception.
+   * @param angleInDegrees  Angle in degrees between 0 and 180.
+   */
   void write(int angleInDegrees);
 }
